# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plume",
    "version": "0.1.0",
    "description": "API REST de Plume. Toutes les requêtes et réponses sont en JSON (sauf mention). Les requêtes qui modifient un état exigent la session et l’en-tête X-CSRF-Token (jeton renvoyé par /auth/login et /auth/session), ainsi qu’un en-tête Origin identique à l’URL publique."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/me/preferences": {
      "get": {
        "operationId": "getMePreferences",
        "summary": "Préférences",
        "tags": [
          "me"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "patch": {
        "operationId": "patchMePreferences",
        "summary": "Modifie les préférences (champs fournis uniquement)",
        "tags": [
          "me"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "theme": {
                    "type": "string",
                    "enum": [
                      "auto",
                      "light",
                      "dark"
                    ]
                  },
                  "accent": {
                    "type": "string",
                    "enum": [
                      "indigo",
                      "blue",
                      "teal",
                      "green",
                      "orange",
                      "rose"
                    ]
                  },
                  "density": {
                    "type": "string",
                    "enum": [
                      "comfortable",
                      "compact"
                    ]
                  },
                  "language": {
                    "type": "string",
                    "enum": [
                      "auto",
                      "fr",
                      "en"
                    ]
                  },
                  "notifications": {
                    "type": "object",
                    "properties": {
                      "desktop": {
                        "type": "boolean"
                      },
                      "sound": {
                        "type": "boolean"
                      },
                      "dailyDigest": {
                        "type": "boolean"
                      }
                    },
                    "additionalProperties": false
                  },
                  "signatureHtml": {
                    "type": "string",
                    "maxLength": 20000
                  },
                  "displayName": {
                    "anyOf": [
                      {
                        "type": "string",
                        "maxLength": 100
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
    "/me/avatar": {
      "get": {
        "operationId": "getMeAvatar",
        "summary": "Photo de profil",
        "tags": [
          "me"
        ],
        "security": [
          {
            "session": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "image/webp": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "operationId": "putMeAvatar",
        "summary": "Remplace la photo (réencodée, sans métadonnées)",
        "tags": [
          "me"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "204": {
            "description": "Succès, sans contenu"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "data": {
                    "type": "string",
                    "minLength": 1,
                    "pattern": "^[A-Za-z0-9+/]*={0,2}$"
                  }
                },
                "required": [
                  "data"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteMeAvatar",
        "summary": "Supprime la photo",
        "tags": [
          "me"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "204": {
            "description": "Succès, sans contenu"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/messages/{id}/invitation/reply": {
      "post": {
        "operationId": "postMessagesIdInvitationReply",
        "summary": "Répond à une invitation (iMIP) : accepter, peut-être ou refuser",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 1500,
              "description": "Identifiant opaque de message"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "accepted",
                      "tentative",
                      "declined"
                    ]
                  }
                },
                "required": [
                  "status"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
    "/messages/{id}/move": {
      "post": {
        "operationId": "postMessagesIdMove",
        "summary": "Déplace le message",
        "tags": [
          "mail"
        ],
        "security": [
          {
            "session": [],
            "csrf": []
          }
        ],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 1500,
              "description": "Identifiant opaque de message"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "folder": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  }
                },
                "required": [
                  "folder"
                ],
                "additionalProperties": false
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-plume_sid"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    },
    "responses": {
      "Error": {
        "description": "Erreur (code stable, sans détail interne)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              },
              "required": [
                "error"
              ]
            }
          }
        }
      }
    }
  }
}